[ip6] implement multiple simultaneous reassembly support for ip6 packages.

+++ src/core/thread/mesh_forwarder.cpp
@@ -792,7 +792,7 @@ start:
             fragmentHeader = reinterpret_cast<Lowpan::FragmentHeader *>(payload);
             fragmentHeader->Init();
             fragmentHeader->SetDatagramSize(aMessage.GetLength());
-            fragmentHeader->SetDatagramTag(aMessage.GetDatagramTag());
+            fragmentHeader->SetDatagramTag(static_cast<uint16_t>(aMessage.GetDatagramTag()));
             fragmentHeader->SetDatagramOffset(0);
 
             payload += fragmentHeader->GetHeaderLength();
@@ -819,7 +819,7 @@ start:
         fragmentHeader = reinterpret_cast<Lowpan::FragmentHeader *>(payload);
         fragmentHeader->Init();
         fragmentHeader->SetDatagramSize(aMessage.GetLength());
-        fragmentHeader->SetDatagramTag(aMessage.GetDatagramTag());
+        fragmentHeader->SetDatagramTag(static_cast<uint16_t>(aMessage.GetDatagramTag()));
         fragmentHeader->SetDatagramOffset(aMessage.GetOffset());
 
         payload += fragmentHeader->GetHeaderLength();
